requirements.txt
allure-pytest==2.13.5
pytest==7.4.3
requests==2.31.0
